Feat: New function constrains enzymes for a protein_draw ecModel
This new function takes a protein_draw ecModel and constrains the upper bounds for model proteins based on experimental data. One feature of this function includes an option to pass as input index matching information between the data and the model, greatly speeding the constraining time necessary for constraining a single model using multiple measurements from the same data set.